Removing queueIn Removing queueOut Allowing standard messages on queue(), not just entity. Promoted set/getDirection from message entity to common method. Promoted direction tests to message trait, which enables coverage on both message classes.
